requirements.txt
numpy<2
faiss-cpu==1.7.4
sentence-transformers==3.0.1
pydantic==2.8.2
python-dotenv==1.0.1
PyMuPDF==1.24.8
Pillow==10.4.0
pytesseract==0.3.10
openai==1.47.0
google-generativeai==0.8.2
groq==0.9.0
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx<0.28
streamlit==1.39.0
